requirements.txt
opencv-python
pillow
umsgpack
urllib3
blend_modes
requests
requests_oauthlib
tornado==4.2
obs-websocket-py
